+++ nary-tree/traversal/preorder_test.go
@@ -4,6 +4,8 @@ import (
 	"fmt"
 	"reflect"
 	"testing"
+
+	"github.com/sko00o/leetcode-adventure/nary-tree/treenode"
 )
 
 func Test_preorder(t *testing.T) {
@@ -18,14 +20,14 @@ func Test_preorder(t *testing.T) {
 		{
 			name: "example 1",
 			args: args{
-				root: exampleTree1,
+				root: treenode.ExampleTree1,
 			},
 			want: []int{1, 3, 5, 6, 2, 4},
 		},
 		{
 			name: "example 2",
 			args: args{
-				root: exampleTree2,
+				root: treenode.ExampleTree2,
 			},
 			want: []int{1, 2, 3, 6, 7, 11, 14, 4, 8, 12, 5, 9, 13, 10},
 		},
